bug fixed
up vote
6
down vote
accepted
This happens when your result is not a result (but a "false" instead).
You should change this line

$sql = 'SELECT * FROM $usertable WHERE PartNumber = $partid';
to this:

$sql = "SELECT * FROM $usertable WHERE PartNumber = $partid";
because the " can interprete $variables while ' cannot.

Works fine with integers (numbers), for strings you need to put the
$variable in single quotes, like

$sql = "SELECT * FROM $usertable WHERE PartNumber = '$partid' ";
If you want / have to work with single quotes, then php CAN NOT
interprete the variables, you will have to do it like this:

 $sql = 'SELECT * FROM '.$usertable.' WHERE string_column =
"'.$string.'" AND integer_column = '.$number.';